refactor play
