+++ lib/tddetective.js
@@ -1,6 +1,7 @@
 'use babel';
 
 import TddetectiveView from './tddetective-view';
+import TddetectiveModel from './tddetective-model';
 import { CompositeDisposable } from 'atom';
 const fs = require('fs')
 var classNames = []
@@ -15,6 +16,7 @@ export default {
 
   activate(state) {
     this.tddetectiveView = new TddetectiveView(state.tddetectiveViewState);
+    this.tddetectiveModel = new TddetectiveModel();
     this.modalPanel = atom.workspace.addModalPanel({
       item: this.tddetectiveView.getElement(),
       visible: false
@@ -47,107 +49,14 @@ export default {
   },
 
   listenToChanges(editor){
-    console.log("listen to changes was called")
+    // console.log("listen to changes was called")
     var editor = editor ? editor : atom.workspace.getActiveTextEditor()
-    console.log(editor)
-    console.log("above was console logged from function call")
+    // console.log(editor)
+    // console.log("above was console logged from function call")
     var self = this
-    editor.onDidChange(function() {
-      classNames = []
-      methodNames = []
-      specFileNames = []
-      self.tddetectiveView.removeAllDecorations(editor);
-      // self.tddetectiveView.removeAllNotifications();
-      self._pickUpChanges(editor);
-    });
-  },
-
-  _pickUpChanges(editor) {
-    var self = this
-    this._findClassAndMethodLine(editor);
-    classNames.forEach(function(aClassName) {
-      var classRange = self._findClassObjRange(aClassName);
-      var className = self._findClassObjName(aClassName);
-      self._manageClassChanges(className, classRange, editor);
-    })
-    methodNames.forEach(function(aMethodName) {
-      var methodRange = self._findMethodObjRange(aMethodName);
-      var methodName = self._findMethodObjName(aMethodName);
-      self._manageMethodChanges(methodName, methodRange, editor)
-    })
-  },
-
-  _findClassAndMethodLine(editor) {
-    var buffer = editor.getBuffer();
-    this._runScanFunction(buffer)
-  },
-
-  _runScanFunction(buffer){
-    buffer.scan(/class /g, function(obj) { classNames.push(obj) })
-    buffer.scan(/def /g, function(obj) { methodNames.push(obj) })
-  },
-
-  _findClassObjRange(aClassName){
-    return classRange = aClassName.computedRange;
-  },
-
-  _findClassObjName(aClassName){
-    return className = aClassName.lineText.substring(6);
-  },
-
-  _findMethodObjRange(aMethodName){
-    return aMethodName.computedRange;
-  },
-
-  _findMethodObjName(aMethodName){
-    var spacePosition = aMethodName.lineText.indexOf("def") + 4
-    var methodName = aMethodName.lineText.substring(spacePosition);
-    if (methodName.indexOf("(") != -1) {
-      var charPosition = methodName.indexOf("(")
-      methodName = methodName.substring(0, charPosition)
-    } else if (methodName.indexOf(" ") != -1) {
-      var charPosition = methodName.indexOf(" ")
-      methodName = methodName.substring(0, charPosition)
-    }
-    return methodName
-  },
-
-  _manageClassChanges(aClassName, classRange, editor) {
-    let selection = this._hasSpecFileName(aClassName)
-    if (!selection) {
-      this.tddetectiveView.updateView(aClassName, classRange, editor)
-    }
-  },
-
-  _hasSpecFileName(aClassName) {
-    var path = this._getSpecPath() + '/' + aClassName.toLowerCase() + '_spec.rb'
-    if (fs.existsSync(path)) {
-      specFileNames.push(path);
-    }
-    return fs.existsSync(path)
-  },
-
-  _getSpecPath() {  //STUBBED IN TEST
-    var root = atom.project.getDirectories()
-    return root[0].lowerCasePath + "/spec"
-  },
-
-  _manageMethodChanges(aMethodName, methodRange, editor) {
-    var selection = this._hasMethodNameInSpecFile(aMethodName);
-    if (!selection) {
-      this.tddetectiveView.updateView(aMethodName, methodRange, editor)
-    }
-  },
-
-  _hasMethodNameInSpecFile(aMethodName) {
-    var searchResult = false
-    specFileNames.forEach(function(specFile) {
-      var fileContent = fs.readFileSync(specFile, "utf-8")
-      if (fileContent.indexOf(aMethodName) != -1) {
-        searchResult = true
-      }
-    })
-    return searchResult
+    console.log(this)
+    console.log("the thing above is THIS")
+    editor.onDidChange(function(){self.tddetectiveModel._onDoneChange(self, editor)});
   }
 
 };
